+++ libs/nx_utils/src/nx/utils/uuid.h
@@ -123,7 +123,7 @@ NX_UTILS_API QString changedGuidByteOrder(const QString& guid);
 using QnUuidSet = QSet<QnUuid>;
 using QnUuidList = QList<QnUuid>;
 
-NX_UTILS_API uint qHash( const QnUuid& uuid, uint seed = 0 ) throw();
+NX_UTILS_API size_t qHash(const QnUuid& uuid, size_t seed = 0 ) throw();
 NX_UTILS_API QDataStream& operator<<(QDataStream& s, const QnUuid& id);
 NX_UTILS_API QDebug operator<<(QDebug dbg, const QnUuid& id);
 NX_UTILS_API QDataStream& operator>>(QDataStream& s, QnUuid& id);
