+++ vms/client/nx_vms_client_desktop/src/nx/vms/client/desktop/analytics/analytics_settings_types.h
@@ -18,7 +18,7 @@ struct DeviceAgentId
     }
 };
 
-inline uint qHash(const DeviceAgentId& key)
+inline size_t qHash(const DeviceAgentId& key)
 {
     return qHash(key.device) + qHash(key.engine);
 }
