+++ vms/libs/appserver2/src/transaction/ubjson_transaction_serializer.h
@@ -110,6 +110,6 @@ namespace ec2
         QCache<CacheKey, const QByteArray> m_cache;
     };
 
-    uint qHash(const QnUbjsonTransactionSerializer::CacheKey &id);
+    size_t qHash(const QnUbjsonTransactionSerializer::CacheKey &id);
 
 }
