+++ vms/libs/nx_vms_update/src/nx/vms/update/package_info.h
@@ -32,7 +32,7 @@ NX_REFLECTION_ENUM_CLASS(Component,
     server
 )
 
-inline uint qHash(Component component, uint seed = 0)
+inline size_t qHash(Component component, size_t seed = 0)
 {
     return ::qHash((uint) (component), seed);
 }
